Improvements to socket and pipe closing API.
This changes the protocol layer, and clarifies the interactions
around close.  The net result is a simpler, and more predictable
flow around close.  It also reduces complexity in the protocols,
eliminates code duplication, and in some cases will reduce lock
contention for applications that make use of short-lived connections.

One of the main benefits is improved code readability and
maintainability.  A few related bugs were found and squashed as
part of this effort.