fixes #123 Survey tests sometimes fail due to state errors
This also represents nearly full test coverage of the surveyor
pattern, including contexts, cancellation, and so forth.  The
respondent also was refactored to eliminate possible races, and
to make it more scalable and apply backpressure correctly.

The surveyor itself now uses a pattern to reduce lock pressure
when running with many contexts in parallel, and also eliminates
an extra data copy of outgoing messages.

While here fixed a few style snafus.