REP protocol fixes and test coverage changes.
Full coverage testing added, and uncovered two problems.  First
the OptionBestEffort option didn't work at all.  Second, closing
a socket did not wake operations on individual contexts.

We also identified a clear opportunity to simplify the receive path
in REP so that the code is much simpler, and eliminates potential
races.  The simpler code performs at about the same as the original.

Finally, we have improved some things in the test support code
for the mock which should make it easier to use and help DNRY.