Substantial transport fixes, including huge test suite overhaul.
This includes a mock for socket operations, and achieves a high level
of coverage for the most important transports.  As part of this, we
have found and fixed a number of bugs across the transport layer.

Note that tuning for KeepAlive and NoDelay has been altered -- all
sockets are now forcibly NoDelay.  Additionally, KeepAlive semantics
more closely match the modern Go net package.