Tighten up test for XSUB and XRESPONDENT.
XSUB in particular no longer attempts to prioritize newer messages.
For the primary use case (devices), there is really little or no
benefit, as the messages are discarded on the send side anyway, and
the receiver really never blocks.  This eliminates a difficult set
of conditions to provide good test coverage for.