fixes #211 Want options for peer credential (IPC)
This is a first work, and includes only support for Linux.  Still
that is likely to meet the needs of most users.  Additional
platform support may be added as we can test, or as the underlying
system libraries add the missing socket options or library calls.

This also simplifies some of the option hanldling for pipes,
and eliminates some potential race conditions.