Req receive timeout timer appears to leak.
A lot of requests fired in succession potentially leave a lot
of timers running around in the background in the client.  This
should fix, although we should also (for performance reasons)
look at replacing the AfterFunc() semantic with a normal timer or
perhaps just using context.Context. (This code predates the existence
of Go standard contexts.)