Go 1.13 is too old.
We need to test with more recent Go versions on Windows and
Darwin.  Additionally, we really would like to matrix test
things on Linux (which runs faster because of more resources
that are available) across a matrix of Go versions.

The matrix currently includes Go 1.13 through Go 1.18.  At
some point in the near future we will be pruning back on
support commitments for older Go versions.